Call MongoClient#refresh instead of #reconnect
It turns out reconnect is aliased to MongoClient#connect, even for
MongoReplicaSetClient. Obviously, this doesn't work at all, and the
calls to reconnect accomplished nothing, except making sure that we
generated a lot of garbage in the logs. From a cursory look, it also
seems like Mongo actually automatically reconnects when the connection
is lost when doing the common operations, so reconnecting does not seem
to be necessary.

In the end, it turns out refresh is the appropriate thing to do in the
event of a MongoReplicaSetClient. For MongoClient, the operation doesn't
do anything, so calling it should be safe.